tests/wpt/tests/html/interaction/focus/focusgroup/tentative/behavior-tokens/listbox.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focusgroup - listbox behavior token defaults</title>
<meta name="assert" content="listbox implies block axis only, no wrap. Explicit wrap adds wrapping.">
<link rel="author" title="Microsoft" href="http://www.microsoft.com/">
<link rel="help" href="https://open-ui.org/components/scoped-focusgroup.explainer/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/shadow-dom/focus-navigation/resources/focus-utils.js"></script>
<script src="../resources/focusgroup-utils.js"></script>

<div focusgroup="listbox">
  <button id=lb1>lb1</button>
  <button id=lb2>lb2</button>
  <button id=lb3>lb3</button>
</div>

<!-- listbox wrap: explicit wrap adds wrapping to non-wrapping default -->
<div focusgroup="listbox wrap">
  <button id=lw1>lw1</button>
  <button id=lw2>lw2</button>
  <button id=lw3>lw3</button>
</div>

<script>

  promise_test(async t => {
    await focusAndKeyPress(document.getElementById('lb1'), kArrowDown);
    assert_equals(document.activeElement.id, 'lb2');
  }, "listbox: ArrowDown navigates to the next item");

  promise_test(async t => {
    await focusAndKeyPress(document.getElementById('lb2'), kArrowUp);
    assert_equals(document.activeElement.id, 'lb1');
  }, "listbox: ArrowUp navigates to the previous item");

  promise_test(async t => {
    await focusAndKeyPress(document.getElementById('lb1'), kArrowRight);
    assert_equals(document.activeElement.id, 'lb1');
  }, "listbox: ArrowRight blocked");

  promise_test(async t => {
    await focusAndKeyPress(document.getElementById('lb1'), kArrowLeft);
    assert_equals(document.activeElement.id, 'lb1');
  }, "listbox: ArrowLeft blocked");

  promise_test(async t => {
    await focusAndKeyPress(document.getElementById('lb3'), kArrowDown);
    assert_equals(document.activeElement.id, 'lb3');
  }, "listbox: ArrowDown does not wrap at end");

  promise_test(async t => {
    await focusAndKeyPress(document.getElementById('lb1'), kArrowUp);
    assert_equals(document.activeElement.id, 'lb1');
  }, "listbox: ArrowUp does not wrap at start");

  // --- listbox wrap: adds wrapping to non-wrapping default ---
  promise_test(async t => {
    await focusAndKeyPress(document.getElementById('lw1'), kArrowDown);
    assert_equals(document.activeElement.id, 'lw2');
  }, "listbox wrap: ArrowDown navigates to the next item");

  promise_test(async t => {
    await focusAndKeyPress(document.getElementById('lw3'), kArrowDown);
    assert_equals(document.activeElement.id, 'lw1');
  }, "listbox wrap: ArrowDown wraps at end");

  promise_test(async t => {
    await focusAndKeyPress(document.getElementById('lw1'), kArrowUp);
    assert_equals(document.activeElement.id, 'lw3');
  }, "listbox wrap: ArrowUp wraps at start");

  promise_test(async t => {
    await focusAndKeyPress(document.getElementById('lw1'), kArrowRight);
    assert_equals(document.activeElement.id, 'lw1');
  }, "listbox wrap: ArrowRight still blocked");

</script>
